% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3pogolem-package.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\title{pokemon}
\format{
A \code{data frame} with 151 observations and 15 variables.
}
\source{
Adapted from \code{highcharter} package.
}
\usage{
pokemon
}
\description{
Statistical information about 151 Pokemon from Nintendo RPG series.
}
\keyword{datasets}
