% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3pogolem-package.R
\docType{data}
\name{pokemon_count}
\alias{pokemon_count}
\alias{pokemon_decile}
\title{pokemon_count}
\format{
A \code{data frame} object with 15 observations and 3 variables.

A \code{data frame} object with 33 observations and 4 variables.
}
\source{
Adapted from \code{highcharter} package.

Adapted from \code{highcharter} package.
}
\usage{
pokemon_count

pokemon_count
}
\description{
A summary of the pokemon table.

A summary of the pokemon table.
}
\keyword{datasets}
