% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3pogolem-package.R
\docType{data}
\name{pokemon_def_vs_att}
\alias{pokemon_def_vs_att}
\title{pokemon_def_vs_att}
\format{
A \code{data frame} object with 15 observations and 5 variables.
}
\source{
Adapted from \code{highcharter} package.
}
\usage{
pokemon_def_vs_att
}
\description{
A summary of the pokemon table.
}
\keyword{datasets}
