% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3pogolem-package.R
\docType{data}
\name{pokemon_density}
\alias{pokemon_density}
\title{pokemon_density
A summary of the pokemon table.}
\format{
A \code{data frame} object with 30 observations and 4 variables.
}
\source{
Adapted from \code{highcharter} package.
}
\usage{
pokemon_density
}
\description{
pokemon_density
A summary of the pokemon table.
}
\keyword{datasets}
